+++ build.gradle.kts
@@ -59,7 +59,7 @@ val guavaVersion = "33.4.0-jre"
 val postgresJdbcVersion = "42.7.5"
 val podamVersion = "8.0.2.RELEASE"
 
-val activitiesVersion = "1.104.0"
+val activitiesVersion = "1.104.1"
 
 dependencies {
   implementation("org.springframework.boot:spring-boot-starter")
