+++ src/main/java/it/gov/pagopa/pu/workflow/wf/exportfile/export/config/ExportFileWFConfig.java
@@ -1,8 +1,8 @@
 package it.gov.pagopa.pu.workflow.wf.exportfile.export.config;
 
 import io.temporal.workflow.Workflow;
+import it.gov.pagopa.payhub.activities.activity.exportflow.ExportFileActivity;
 import it.gov.pagopa.payhub.activities.activity.exportflow.UpdateExportFileStatusActivity;
-import it.gov.pagopa.payhub.activities.activity.exportflow.debtposition.ExportFileActivity;
 import it.gov.pagopa.pu.workflow.config.temporal.BaseWfConfig;
 import it.gov.pagopa.pu.workflow.config.temporal.TemporalWFImplementationCustomizer;
 import it.gov.pagopa.pu.workflow.wf.exportfile.export.activity.ScheduleExportFileExpirationActivity;
