+++ src/test/java/it/gov/pagopa/pu/workflow/wf/debtposition/sync/BaseDPSynchronizeWFTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.workflow.wf.debtposition.sync;
 
-import it.gov.pagopa.payhub.activities.activity.debtposition.FinalizeDebtPositionSyncStatusActivity;
+import it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.FinalizeDebtPositionSyncStatusActivity;
 import it.gov.pagopa.payhub.activities.activity.debtposition.ionotification.IONotificationDebtPositionActivity;
 import it.gov.pagopa.payhub.activities.dto.debtposition.DebtPositionIoNotificationDTO;
 import it.gov.pagopa.payhub.activities.dto.debtposition.syncwfconfig.GenericWfExecutionConfig;
