+++ src/test/java/it/gov/pagopa/pu/workflow/wf/debtposition/sync/wf_async_gpd/SynchronizeAsyncGpdWFTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.workflow.wf.debtposition.sync.wf_async_gpd;
 
-import it.gov.pagopa.payhub.activities.activity.debtposition.gpd.SynchronizeInstallmentGpdActivity;
+import it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.gpd.SynchronizeInstallmentGpdActivity;
 import it.gov.pagopa.payhub.activities.dto.debtposition.syncwfconfig.GenericWfExecutionConfig;
 import it.gov.pagopa.pu.debtposition.dto.generated.DebtPositionDTO;
 import it.gov.pagopa.pu.workflow.dto.PaymentEventRequestDTO;
