+++ src/test/java/it/gov/pagopa/pu/workflow/wf/debtposition/sync/wf_sync_aca/SynchronizeSyncAcaWFTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.workflow.wf.debtposition.sync.wf_sync_aca;
 
-import it.gov.pagopa.payhub.activities.activity.debtposition.aca.SynchronizeInstallmentAcaActivity;
+import it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.aca.SynchronizeInstallmentAcaActivity;
 import it.gov.pagopa.payhub.activities.dto.debtposition.syncwfconfig.GenericWfExecutionConfig;
 import it.gov.pagopa.pu.debtposition.dto.generated.DebtPositionDTO;
 import it.gov.pagopa.pu.workflow.dto.PaymentEventRequestDTO;
