+++ src/test/java/it/gov/pagopa/pu/workflow/wf/debtposition/sync/wf_sync_gpdpreload/SynchronizeSyncGpdPreLoadWFTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.workflow.wf.debtposition.sync.wf_sync_gpdpreload;
 
-import it.gov.pagopa.payhub.activities.activity.debtposition.gpdpreload.SynchronizeInstallmentGpdPreLoadActivity;
+import it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.gpdpreload.SynchronizeInstallmentGpdPreLoadActivity;
 import it.gov.pagopa.payhub.activities.dto.debtposition.syncwfconfig.GenericWfExecutionConfig;
 import it.gov.pagopa.pu.debtposition.dto.generated.DebtPositionDTO;
 import it.gov.pagopa.pu.workflow.dto.PaymentEventRequestDTO;
